+++ scripts/macros/monsterFeatures/bat/bat.js
@@ -1,0 +1,4 @@
+import {keenHearing} from './keenHearing.js';
+export let bat = {
+    'keenHearing': keenHearing
+};