+++ scripts/macros/monsterFeatures/bat/keenHearing.js
@@ -1,0 +1,3 @@
+export function keenHearing(skillId, options) {
+    return skillId != 'prc' ? false : {'label': 'This check relies hearing.', 'type': 'advantage'};
+}