+++ scripts/macros/monsterFeatures/cat/cat.js
@@ -1,0 +1,4 @@
+import {keenSmell} from './keenSmell.js';
+export let cat = {
+    'keenSmell': keenSmell
+};