+++ scripts/macros/monsterFeatures/cat/keenSmell.js
@@ -1,0 +1,3 @@
+export function keenSmell(skillId, options) {
+    return skillId != 'prc' ? false : {'label': 'This check relies on smell.', 'type': 'advantage'};
+}