+++ scripts/macros/monsterFeatures/eagle/eagle.js
@@ -1,0 +1,4 @@
+import {eyesOfTheEagle} from '../../items/eyesOfTheEagle.js';
+export let eagle = {
+    'keenSight': eyesOfTheEagle
+};