+++ scripts/macros/monsterFeatures/owl/keenHearingAndSight.js
@@ -1,0 +1,3 @@
+export function keenHearingAndSight(skillId, options) {
+    return skillId != 'prc' ? false : {'label': 'This check relies hearing or sight.', 'type': 'advantage'};
+}