+++ scripts/macros/monsterFeatures/owl/owl.js
@@ -1,0 +1,4 @@
+import {keenHearingAndSight} from './keenHearingAndSight.js';
+export let owl = {
+    'keenHearingAndSight': keenHearingAndSight
+};