+++ scripts/macros/monsterFeatures/pseudodragon/keenSenses.js
@@ -1,0 +1,3 @@
+export function keenSenses(skillId, options) {
+    return skillId != 'prc' ? false : {'label': 'This check relies sight, hearing, or small.', 'type': 'advantage'};
+}