+++ scripts/macros/monsterFeatures/pseudodragon/pseudodragon.js
@@ -1,0 +1,4 @@
+import {keenSenses} from './keenSenses.js';
+export let pseudodragon = {
+    'keenSenses': keenSenses
+};