+++ scripts/macros/monsterFeatures/vulture/keenSightAndSmell.js
@@ -1,0 +1,3 @@
+export function keenSightAndSmell(skillId, options) {
+    return skillId != 'prc' ? false : {'label': 'This check relies sight or smell.', 'type': 'advantage'};
+}