+++ scripts/macros/monsterFeatures/vulture/vulture.js
@@ -1,0 +1,4 @@
+import {keenSightAndSmell} from './keenSightAndSmell.js';
+export let vulture = {
+    'keenSightAndSmell': keenSightAndSmell
+};