+++ scripts/macros/monsterFeatures/wolf/keenHearingAndSmell.js
@@ -1,0 +1,3 @@
+export function keenHearingAndSmell(skillId, options) {
+    return skillId != 'prc' ? false : {'label': 'This check relies hearing or smell.', 'type': 'advantage'};
+}