+++ scripts/macros/monsterFeatures/wolf/wolf.js
@@ -1,0 +1,4 @@
+import {keenHearingAndSmell} from './keenHearingAndSmell.js';
+export let wolf = {
+    'keenHearingAndSmell': keenHearingAndSmell
+};